Define $RM for make -R
When make is invoked as `make -R`, then `$(RM)` is not defined.  It is
even worse if the Makefile includes `MAKEFLAGS += R`, which apparently
causes `$(RM)` to be cleared during processing, after the file is
processed even.

In my testing, I found that assigning `$(RMDIR)` using `:=` is enough
when using `MAKEFLAGS` and assigning `$(RM)` to a default value is
necessary when `-R` appears on the command line.  Neither is sufficient
on its own.