Use test package without an underscore in name
This test suite is not the right place to deal with undescore<>hyphen
conversion drama.

Apparently, pip (or setuptools or something) previously did automatic
conversion between dashes and underscores in package names depending on
situation. Failing CI builds indicate that this behavior has changed.
We are not testing pip here, so we switch to a test package with a simpler
name.